+++ release/components/body/body-cell.component.js
@@ -93,6 +93,13 @@ var DataTableBodyCellComponent = (function () {
         enumerable: true,
         configurable: true
     });
+    Object.defineProperty(DataTableBodyCellComponent.prototype, "valueStripped", {
+        get: function () {
+            return this.stripHtml(this.value);
+        },
+        enumerable: true,
+        configurable: true
+    });
     DataTableBodyCellComponent.prototype.ngOnDestroy = function () {
         if (this.cellTemplate) {
             this.cellTemplate.clear();
@@ -165,10 +172,15 @@ var DataTableBodyCellComponent = (function () {
         if (sort)
             return sort.dir;
     };
+    DataTableBodyCellComponent.prototype.stripHtml = function (html) {
+        if (!html.replace)
+            return html;
+        return html.replace(/<\/?[^>]+(>|$)/g, '');
+    };
     DataTableBodyCellComponent.decorators = [
         { type: core_1.Component, args: [{
                     selector: 'datatable-body-cell',
-                    template: "\n    <div class=\"datatable-body-cell-label\">\n      <label\n        *ngIf=\"column.checkboxable\" \n        class=\"datatable-checkbox\">\n        <input \n          type=\"checkbox\"\n          [checked]=\"isSelected\"\n          (click)=\"onCheckboxChange($event)\" \n        />\n      </label>\n      <span\n        *ngIf=\"!column.cellTemplate\"\n        [title]=\"value\"\n        [innerHTML]=\"value\">\n      </span>\n      <ng-template #cellTemplate\n        *ngIf=\"column.cellTemplate\"\n        [ngTemplateOutlet]=\"column.cellTemplate\"\n        [ngOutletContext]=\"{\n          value: value,\n          row: row,\n          column: column,\n          isSelected: isSelected,\n          onCheckboxChangeFn: onCheckboxChangeFn,\n          activateFn: activateFn\n        }\">\n      </ng-template>\n    </div>\n  ",
+                    template: "\n    <div class=\"datatable-body-cell-label\">\n      <label\n        *ngIf=\"column.checkboxable\"\n        class=\"datatable-checkbox\">\n        <input\n          type=\"checkbox\"\n          [checked]=\"isSelected\"\n          (click)=\"onCheckboxChange($event)\"\n        />\n      </label>\n      <span\n        *ngIf=\"!column.cellTemplate\"\n        [title]=\"valueStripped\"\n        [innerHTML]=\"value\">\n      </span>\n      <ng-template #cellTemplate\n        *ngIf=\"column.cellTemplate\"\n        [ngTemplateOutlet]=\"column.cellTemplate\"\n        [ngOutletContext]=\"{\n          value: value,\n          row: row,\n          column: column,\n          expanded: expanded,\n          isSelected: isSelected,\n          onCheckboxChangeFn: onCheckboxChangeFn,\n          activateFn: activateFn\n        }\">\n      </ng-template>\n    </div>\n  ",
                     host: {
                         class: 'datatable-body-cell'
                     }
@@ -183,6 +195,7 @@ var DataTableBodyCellComponent = (function () {
         'column': [{ type: core_1.Input },],
         'rowHeight': [{ type: core_1.Input },],
         'isSelected': [{ type: core_1.Input },],
+        'expanded': [{ type: core_1.Input },],
         'sorts': [{ type: core_1.Input },],
         'activate': [{ type: core_1.Output },],
         'cellTemplate': [{ type: core_1.ViewChild, args: ['cellTemplate', { read: core_1.ViewContainerRef },] },],