+++ release/components/body/body-row.component.d.ts
@@ -4,6 +4,7 @@ export declare class DataTableBodyRowComponent {
     private scrollbarHelper;
     columns: any[];
     innerWidth: number;
+    expanded: boolean;
     rowClass: any;
     row: any;
     offsetX: number;