+++ release/events.d.ts
@@ -1,2 +1,2 @@
-export declare const MouseEvent: MouseEvent;
-export declare const KeyboardEvent: KeyboardEvent;
+export declare const mouseEvent: MouseEvent;
+export declare const keyboardEvent: KeyboardEvent;