+++ release/events.js
@@ -1,5 +1,5 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.MouseEvent = global.MouseEvent;
-exports.KeyboardEvent = global.KeyboardEvent;
+exports.mouseEvent = global.MouseEvent;
+exports.keyboardEvent = global.KeyboardEvent;
 //# sourceMappingURL=events.js.map