+++ docs/changelog.md
@@ -1,6 +1,12 @@
 # Changelog
+## 10.0.0
+- BREAKING: `$$expanded` is no longer on row of detail templates, use `let-expanded` to get pop now
+- Feature: Always center activated page on pager
+- Bug: Remove row mutations of `$$index` and `$$expanded`
+- Bug: Pagination broken on large number of rows
+- Bug: Strip HTML from title tags (#892)
 
-# 9.3.1
+## 9.3.1
 - Bug: More Universal Rendering Fixes (#795)
 - Chore: Build optimizations from new webpack version